Refactor code in different directories
