+++ index.ts
@@ -3,7 +3,7 @@ import express from "express";
 import cors from "cors";
 import { Server } from "colyseus";
 import { monitor } from "@colyseus/monitor";
-import { PandemicTogetherRoom } from "./src/PandemicTogetherRoom";
+import { PandemicTogetherRoom } from "./src/rooms/PandemicTogetherRoom";
 
 const port = Number(process.env.PORT || 2567) + Number(process.env.NODE_APP_INSTANCE || 0);
 const app = express()
