+++ src/rooms/PandemicTogetherRoom.ts
@@ -1,49 +1,11 @@
 import { Room, Client } from "colyseus";
 import { Schema, ArraySchema, MapSchema, type } from "@colyseus/schema";
-import { GameElements } from './GameElements'
-import * as Constants from './constants';
-import * as DeckFunctions from './DeckFunctions';
+import { GameElements } from "../game-components/GameElements"
+import * as Constants from "../common/constants";
+import * as DeckFunctions from "../game-components/DeckFunctions";
+import Card from "../game-components/Card";
+import Player from "../game-components/Player";
 
-class Card extends Schema {
-  @type("string") cardId: string;
-  @type("string") elementId: string;
-  @type("string") type: string;
-  @type("string") name: string;
-  @type("string") description: string;
-  @type("string") action: string;
-  @type("number") tokens: number;
-  @type("number") maxImpactPerElement: number;
-  @type(["string"]) impactedElements = new ArraySchema<string>();
-  @type("string") cardHolder: string;
-  @type("boolean") contained: boolean;
-
-  constructor({cardId, elementId, type, name, description, action, maxImpactPerElement = 0, impactedElements = []}) {
-    super();
-    this.cardId = cardId;
-    this.elementId = elementId;
-    this.type = type;
-    this.name = name;
-    this.description = description;
-    this.action = action;
-    this.tokens = (type == Constants.CARD_TYPE_VIRUS)? 1 : 0;
-    this.maxImpactPerElement = maxImpactPerElement;
-    this.contained = false;
-
-    for (var i = 0; i < impactedElements.length; i++) {
-      this.impactedElements[i] = impactedElements[i];
-    }
-  }
-}
-
-class Player extends Schema {
-  @type("string") sessionId: string;
-  @type("string") name: string;
-  @type("boolean") gameStart: boolean; //TODO when player clicks start button
-  @type([ Card ]) hand = new ArraySchema<Card>();
-  @type([ Card ]) advantages = new ArraySchema<Card>();
-  @type([ Card ]) disadvantages = new ArraySchema<Card>();
-  @type([ Card ]) virusField = new ArraySchema<Card>();
-}
 
 class State extends Schema {
   @type("string") gameState: string;
@@ -56,7 +18,6 @@ class State extends Schema {
   @type([ Card ]) disadvantagesDeck = new ArraySchema<Card>();
   @type([ Card ]) advantagesDeck = new ArraySchema<Card>();
 }
-
 
 export class PandemicTogetherRoom extends Room {
